+++ spec/unit/provider/mount/mount_spec.rb
@@ -347,7 +347,16 @@ describe Chef::Provider::Mount::Mount do
       it "should use mount -o remount if remount is supported" do
         @new_resource.supports({:remount => true})
         @current_resource.mounted(true)
-        @provider.should_receive(:shell_out!).with("mount -o remount #{@new_resource.mount_point}")
+        @provider.should_receive(:shell_out!).with("mount -o remount,defaults #{@new_resource.mount_point}")
+        @provider.remount_fs
+      end
+
+      it "should use mount -o remount with new mount options if remount is supported" do
+        @new_resource.supports({:remount => true})
+        options = "rw,noexec,noauto"
+        @new_resource.options(%w{rw noexec noauto})
+        @current_resource.mounted(true)
+        @provider.should_receive(:shell_out!).with("mount -o remount,rw,noexec,noauto #{@new_resource.mount_point}")
         @provider.remount_fs
       end